requirements.txt
pathlib==1.0.1
python-decouple==3.8
python-telegram-bot==20.5
